feat: add security and access schemas, including permissions, roles, and configurations refactor: reorganize system behavior schemas and introduce data retention and module toggle configurations chore: update .gitignore to exclude GitHub workflows
